+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPTMR87VMJKF124ZH1DKVV7R"
+        "01KPTVWFZ2PD4CV5KXGMF9H641"
     )
     r"""The unique identifier of the data source"""