+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPTMR87RHZS7268CCX5FFDA2"
+        "01KPTVWFZ0RR4D1DTZ6AAQEW1F"
     )
     r"""The unique identifier of the data source"""