+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPTMR87TA8K8HXG9QQ36Y6HR"
+        "01KPTVWFZ156QA2EHZB03E6C3M"
     )
     r"""The unique identifier of the data source"""