+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPTMR87XFC6NQEYGKRRF85N9"
+        "01KPTVWFZ58VZ46FP0NQAD07PR"
     )
     r"""The unique identifier of the data source"""