+++ packages/orq-rc/RELEASES.md
@@ -7249,3 +7249,13 @@ Based on:
 - [python v4.8.0-rc.48] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.48 - packages/orq-rc
+
+## 2026-04-22 15:13:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.49] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.49 - packages/orq-rc